+++ src/main/java/cubicchunks/asm/mixin/core/common/MixinWorld.java
@@ -29,11 +29,11 @@ import cubicchunks.ICubicChunksWorldType;
 import cubicchunks.lighting.LightingManager;
 import cubicchunks.util.AddressTools;
 import cubicchunks.util.CubeCoords;
-import cubicchunks.world.IColumnProvider;
-import cubicchunks.world.ICubeCache;
 import cubicchunks.world.ICubicWorld;
 import cubicchunks.world.NotCubicChunksWorldException;
 import cubicchunks.world.cube.Cube;
+import cubicchunks.world.provider.IColumnProvider;
+import cubicchunks.world.provider.ICubeCache;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
