+++ src/main/java/cubicchunks/debug/item/RelightSkyBlockItem.java
@@ -27,8 +27,8 @@ import cubicchunks.debug.ItemRegistered;
 import cubicchunks.network.PacketCube;
 import cubicchunks.network.PacketDispatcher;
 import cubicchunks.util.CubeCoords;
-import cubicchunks.world.ICubeCache;
 import cubicchunks.world.ICubicWorld;
+import cubicchunks.world.provider.ICubeCache;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
