+++ src/main/java/cubicchunks/network/ClientHandler.java
@@ -25,15 +25,15 @@ package cubicchunks.network;
 
 import com.google.common.base.Throwables;
 import cubicchunks.CubicChunks;
-import cubicchunks.client.ClientCubeCache;
 import cubicchunks.lighting.LightingManager;
 import cubicchunks.util.CubeCoords;
 import cubicchunks.world.ClientOpacityIndex;
-import cubicchunks.world.IColumnProvider;
 import cubicchunks.world.ICubicWorldClient;
 import cubicchunks.world.column.Column;
 import cubicchunks.world.cube.BlankCube;
 import cubicchunks.world.cube.Cube;
+import cubicchunks.world.provider.ClientCubeCache;
+import cubicchunks.world.provider.IColumnProvider;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.client.Minecraft;
 import net.minecraft.nbt.NBTTagCompound;
