+++ src/main/java/cubicchunks/server/chunkio/IONbtWriter.java
@@ -25,7 +25,6 @@ package cubicchunks.server.chunkio;
 
 import cubicchunks.CubicChunks;
 import cubicchunks.util.Coords;
-import cubicchunks.world.ChunkSectionHelper;
 import cubicchunks.world.OpacityIndex;
 import cubicchunks.world.column.Column;
 import cubicchunks.world.cube.Cube;
