+++ src/main/java/cubicchunks/util/processor/QueueProcessor.java
@@ -25,7 +25,8 @@ package cubicchunks.util.processor;
 
 import com.google.common.collect.Sets;
 import cubicchunks.util.ArrayBatchedQueue;
-import cubicchunks.world.IColumnProvider;
+import cubicchunks.world.provider.IColumnProvider;
+
 import java.util.Set;
 
 public abstract class QueueProcessor<T> {
